@@ -1,12 +1,7 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import Card from '@material-ui/core/Card';
+import { makeStyles } from '@mui/styles';
+import {Card, CardActions, CardContent, CardMedia, Chip, Typography} from '@mui/material';
 //import CardActionArea from '@material-ui/core/CardActionArea';
-import CardActions from '@material-ui/core/CardActions';
-import CardContent from '@material-ui/core/CardContent';
-import CardMedia from '@material-ui/core/CardMedia';
-import Chip from "@material-ui/core/Chip";
-import Typography from '@material-ui/core/Typography';
 
 const useStyles = makeStyles({
     root: {
