@@ -1,5 +1,5 @@
 import React from 'react';
-import {Typography} from "@material-ui/core";
+import {Typography} from "@mui/material";
 const P404 = props => {
     //const {classes} = props;
     return <div>
